SKVR master data version 3.8.2024
New: 1) The data from Scandinavian areas (metsäsuomalaiset) and Tveri Karelia has been mapped to the level of parishes (or equivalent administrative units). 2)The place name and collector lists are updated: only one list that includes the attributes both for SKVR and JR, all the attributes beginning with skvr. 3)The collections of Samuli and Jenni Paulaharju are separated from each other (in the archives these still are one collection attributed to Samuli Paulaharju).